requirements.txt
numpy
flask
python-docx
transformers
pymupdf
sentence_transformers
faiss-cpu
boto3